[Data] Fix `PipelineIterator.to_tf` (#32530)
We move to_tf logic to DatasetIterator, leveraging the respective iter_batches implementations for BulkDatasetIterator and PipelineDatasetIterator.
